refactor: exit boot services and capture framebuffer
Instead of drawing to the display using UEFI boot services, we now exit
the boot services as soon as possible after initializing a `Display`
struct, abstracting around the captured framebuffer pointer. The main
challenge was being able to wait between individual frame renders to
preserve a constant frame rate, since we no longer had access to
`stall`, which was a part of boot services. To get around this,
`ApicTimer` was implemented.

This is now more convenient to do after migrating to the latest v0.36
`uefi` crate, that now maintains a global state for the boot services
instead of requiring to be passed around.

Doing this gives a lot more flexibility and opens up the possibility of
being able to implement an audio driver in the future along with many
more potential improvements.