feat: render directly from video frames
Formerly, we used to convert video frames into an intermediate ASCII
format that was easy to parse, and render the video from that during
runtime. That was not a very... elegant approach. Instead, we now embed
the frames themselves into the binary at build time and decode them
using `zune_png` (which was conveniently `no_std` compatible) at
runtime.

The biggest benefit of this approach is being able to do true image
processing, such as scaling and color manipulation. As a result, we can
now scale the frames to the entire display instead of being limited to
the size of the raw frames.

Some other important changes that were made:
* `ApicTimer` calibration being faulty during frequency estimation,
  since the divisor was not correctly being set before the timing check
* Removed `White` as an enum varient for `Color` and made it a constant,
  alongside `Color::BLACK`
* Various utility methods for `Color`: `to_grayscale`, `to_bw`,
  `to_two_tone`, `invert`
* Implemented `Into<(u8, u8, u8)>` (RGB format) for `Color`
* `resize!` macro for convenient way to resize images given a color
  space using the `resize` crate (UNUSED)
* Remove redundant doc comment for `TimeExt` implementation for
  `uefi::runtime::Time`